Add files via upload
This Jupyter notebook is for implementing the trained neural model to fit the Generalized Extreme Value distribution to univariate extremes. This demonstration is for the Pre-Industrial scenario only using the Case Study data and CCSM3 data mentioned in the manuscript. We cannot make the case study data available, but this will provide you with a perspective to apply it to real data along with bootstrapping.